Correction de tous les problemes

+++ include/Server.hpp
@@ -19,9 +19,9 @@ class Server
 		~Server();
 		Server(const Server& copy);
 		Server&	operator=(const Server& copy);
-		int getPort() const;
-		void setClientBodyBufferSize(int size);
-		int  getClientBodyBufferSize() const ;
+		int		getPort() const;
+		void	setClientBodyBufferSize(int size);
+		size_t	getClientBodyBufferSize() const ;
 		const std::string &getServerName() const;
 		const std::string &getRoot() const;
 		const std::vector<Location> &getLocations() const;
